网关： 1、修改反馈节点节点channel错误问题
节点：
1、修改由于上传时间大于采集时间，重置了定时器问题

+++ app/radio_protocal.c
@@ -2,7 +2,7 @@
 * @Author: zxt
 * @Date:   2017-12-26 16:36:20
 * @Last Modified by:   zxt
-* @Last Modified time: 2018-10-19 15:42:47
+* @Last Modified time: 2018-10-22 19:06:41
 */
 #include "../general.h"
 
@@ -422,7 +422,7 @@ void ConcenterProtocalDispath(EasyLink_RxPacket * protocalRxPacket)
 			if(ConcenterReadSynTimeFlag())
 			{
 #ifdef SUPPORT_STRATEGY_SORT
-				ConcenterSetNodeChannel(bufTemp->srcAddr, ((uint16_t)bufTemp->load[0])<<8 + bufTemp->load[1]);
+				ConcenterSetNodeChannel(bufTemp->srcAddr, (((uint32_t)(bufTemp->load[0]))<<8) + (uint32_t)(bufTemp->load[1]));
 #endif // SUPPORT_STRATEGY_SORT				
 				ConcenterRadioSendSynTime(bufTemp->dstAddr, bufTemp->srcAddr);
 			}
